refact: MemberCommand / Query 구조 리팩토링
