+++ src/main/java/com/midas/shootpointer/domain/member/controller/query/MemberQueryController.java
@@ -1,0 +1,67 @@
+package com.midas.shootpointer.domain.member.controller.query;
+
+import com.midas.shootpointer.domain.highlight.repository.HighlightQueryRepository;
+import com.midas.shootpointer.domain.member.dto.MemberResponseDto;
+import com.midas.shootpointer.domain.member.entity.Member;
+import com.midas.shootpointer.domain.memberbacknumber.repository.MemberBackNumberRepository;
+import com.midas.shootpointer.global.annotation.CustomLog;
+import com.midas.shootpointer.global.security.SecurityUtils;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.media.Content;
+import io.swagger.v3.oas.annotations.media.Schema;
+import io.swagger.v3.oas.annotations.tags.Tag;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/member")
+@Tag(name = "회원 관리", description = "회원 관리 API")
+public class MemberQueryController {
+	
+	private final MemberBackNumberRepository memberBackNumberRepository;
+	private final HighlightQueryRepository highlightQueryRepository;
+	
+	@CustomLog("회원 정보 조회")
+	@Operation(
+		summary = "회원 정보 조회 API - [담당자 : 박재성]",
+		responses = {
+			@io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "회원 정보 조회 성공",
+				content = @Content(mediaType = "application/json",
+					schema = @Schema(implementation = MemberResponseDto.class))),
+			@io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "4XX", description = "회원 정보 조회 실패",
+				content = @Content(mediaType = "application/json",
+					schema = @Schema(implementation = com.midas.shootpointer.global.dto.ApiResponse.class)))
+		}
+	)
+	@GetMapping("/me")
+	public MemberResponseDto getCurrentMember() {
+		Member currentMember = SecurityUtils.getCurrentMember();
+		UUID memberId = currentMember.getMemberId();
+		
+		// 등번호 조회
+		Integer backNumber = memberBackNumberRepository.findByMemberId(memberId)
+			.map(memberBackNumber -> memberBackNumber.getBackNumber().getBackNumber().getNumber())
+			.orElse(0);
+		
+		// 2점, 3점슛 조회
+		Integer totalTwoPoint = highlightQueryRepository.sumTwoPointByMemberId(memberId);
+		Integer totalThreePoint = highlightQueryRepository.sumThreePointByMemberId(memberId);
+		
+		// 하이라이트 개수 조회
+		Integer highlightCount = highlightQueryRepository.countByMemberId(memberId);
+		
+		return MemberResponseDto.builder()
+			.memberId(currentMember.getMemberId())
+			.email(currentMember.getEmail())
+			.username(currentMember.getUsername())
+			.backNumber(backNumber)
+			.totalTwoPoint(totalTwoPoint)
+			.totalThreePoint(totalThreePoint)
+			.highlightCount(highlightCount)
+			.build();
+	}
+}
